@@ -17,10 +17,10 @@ func main() {
 		var err error
 		var id int
 
-		if data.Api == true {
-			id, err = CheckApi(data, *c)
+		if data.Api {
+			id, err = CheckApi(data, c)
 		} else {
-			id, err = Signin(data.Email, *c, false)
+			id, err = Signin(data.Email, c, false)
 		}
 
 		if err != nil {
@@ -34,7 +34,7 @@ func main() {
 		var data auth
 		c.Bind(&data)
 
-		if err := CheckCode(data.ID, data.Code, *c); err != nil {
+		if err := CheckCode(data.ID, data.Code, c); err != nil {
 			c.JSON(401, gin.H{"error": err.Error()})
 		} else {
 			c.JSON(200, gin.H{"id": data.ID})
@@ -42,14 +42,14 @@ func main() {
 	})
 
 	r.PUT("/logout", Auth(), func(c *gin.Context) {
-		Logout(*c)
+		Logout(c)
 	})
 
 	r.GET("/profile", func(c *gin.Context) {
 		var data user
 		c.Bind(&data)
 
-		target := GetTarget(data.ID, *c)
+		target := GetTarget(data.ID, c)
 		if user, err := GetProfile(data.ID); err != nil {
 			c.JSON(404, gin.H{"error": err.Error()})
 		} else {
@@ -61,7 +61,7 @@ func main() {
 		var data user
 		c.Bind(&data)
 
-		ChangeOnline(data.Online, *c)
+		ChangeOnline(data.Online, c)
 	})
 
 	r.GET("/country", func(c *gin.Context) {
@@ -80,7 +80,7 @@ func main() {
 		var data user
 		c.BindJSON(&data)
 
-		if err := Change(data, *c); err != nil {
+		if err := Change(data, c); err != nil {
 			c.JSON(400, gin.H{"error": err.Error()})
 		} else {
 			c.JSON(200, gin.H{})
@@ -97,16 +97,16 @@ func main() {
 	})
 
 	r.POST("/templates", Auth(), func(c *gin.Context) {
-		CreateTemplates(c.PostForm("text"), *c)
+		CreateTemplates(c.PostForm("text"), c)
 	})
 
 	r.GET("/templates", Auth(), func(c *gin.Context) {
-		text := GetTemplates(*c)
+		text := GetTemplates(c)
 		c.JSON(200, text)
 	})
 
 	r.PUT("/deactivate", Auth(), func(c *gin.Context) {
-		DeactivateAccount(*c)
+		DeactivateAccount(c)
 	})
 
 	r.POST("/getUsers", func(c *gin.Context) {
@@ -128,49 +128,49 @@ func main() {
 		var data target
 		c.Bind(&data)
 
-		AddLike(data, *c)
+		AddLike(data, c)
 	})
 
 	r.POST("/private", Auth(), func(c *gin.Context) {
 		var data target
 		c.Bind(&data)
 
-		AddPrivate(data.Target, *c)
+		AddPrivate(data.Target, c)
 	})
 
 	r.POST("/access", Auth(), func(c *gin.Context) {
 		var data target
 		c.Bind(&data)
 
-		AddAccess(data.Target, *c)
+		AddAccess(data.Target, c)
 	})
 
 	r.DELETE("/like", Auth(), func(c *gin.Context) {
 		var data target
 		c.Bind(&data)
 
-		DeleteLike(data.Target, *c)
+		DeleteLike(data.Target, c)
 	})
 
 	r.DELETE("/private", Auth(), func(c *gin.Context) {
 		var data target
 		c.Bind(&data)
 
-		DeletePrivate(data.Target, *c)
+		DeletePrivate(data.Target, c)
 	})
 
 	r.DELETE("/access", Auth(), func(c *gin.Context) {
 		var data target
 		c.Bind(&data)
 
-		DeleteAccess(data.Target, *c)
+		DeleteAccess(data.Target, c)
 	})
 
 	r.POST("/targets", Auth(), func(c *gin.Context) {
 		var data target
 		c.BindJSON(&data)
 
-		count, res := GetTargets(data, *c)
+		count, res := GetTargets(data, c)
 		c.JSON(200, gin.H{
 			"data":  res,
 			"count": count,
@@ -178,12 +178,12 @@ func main() {
 	})
 
 	r.GET("/notifications", Auth(), func(c *gin.Context) {
-		res := GetNotifications(*c)
+		res := GetNotifications(c)
 		c.JSON(200, res)
 	})
 
 	r.POST("/upload", Auth(), func(c *gin.Context) {
-		if err := UploadImage(c.PostForm("dir"), *c); err != nil {
+		if err := UploadImage(c.PostForm("dir"), c); err != nil {
 			c.JSON(400, gin.H{"error": err.Error()})
 		} else {
 			c.JSON(200, gin.H{})
@@ -191,17 +191,17 @@ func main() {
 	})
 
 	r.PUT("/changeImageDir", Auth(), func(c *gin.Context) {
-		ChangeImageDir(c.Query("isAvatar"), c.Query("dir"), c.Query("new"), c.Query("number"), *c)
+		ChangeImageDir(c.Query("isAvatar"), c.Query("dir"), c.Query("new"), c.Query("number"), c)
 		c.JSON(200, "chanched")
 	})
 
 	r.PUT("/replaceAvatar", Auth(), func(c *gin.Context) {
-		ReplaceAvatar(c.Query("dir"), c.Query("number"), *c)
+		ReplaceAvatar(c.Query("dir"), c.Query("number"), c)
 		c.JSON(200, "replaced")
 	})
 
 	r.DELETE("/deleteImage", Auth(), func(c *gin.Context) {
-		DeleteImage(c.Query("isAvatar"), c.Query("dir"), c.Query("number"), *c)
+		DeleteImage(c.Query("isAvatar"), c.Query("dir"), c.Query("number"), c)
 		c.JSON(200, "deleted")
 	})
 
@@ -216,14 +216,14 @@ func main() {
 	})
 
 	r.GET("/chat", Auth(), func(c *gin.Context) {
-		Chat(c.Writer, c.Request, *c)
+		Chat(c.Writer, c.Request, c)
 	})
 
 	r.POST("/getRooms", Auth(), func(c *gin.Context) {
 		var data rooms
 		c.BindJSON(&data)
 
-		res, ids := GetRooms(data, *c)
+		res, ids := GetRooms(data, c)
 		c.JSON(200, gin.H{
 			"rooms": res["rooms"],
 			"users": res["users"],
@@ -243,12 +243,12 @@ func main() {
 		var data messages
 		c.Bind(&data)
 
-		res := GetMessages(data, *c)
+		res := GetMessages(data, c)
 		c.JSON(200, res)
 	})
 
 	r.GET("/getParamsAfterLogin", Auth(), func(c *gin.Context) {
-		user, messages := GetParamsAfterLogin(*c)
+		user, messages := GetParamsAfterLogin(c)
 		c.JSON(200, gin.H{
 			"user":     user,
 			"messages": messages,
@@ -266,7 +266,7 @@ func main() {
 	})
 
 	r.PUT("/trial", Auth(), func(c *gin.Context) {
-		expires, err := TriggerTrial(*c)
+		expires, err := TriggerTrial(c)
 		c.JSON(200, gin.H{
 			"err":     err,
 			"expires": expires,
@@ -274,7 +274,7 @@ func main() {
 	})
 
 	r.GET("/checkPremium", Auth(), func(c *gin.Context) {
-		left := CheckPremium(*c)
+		left := CheckPremium(c)
 		c.JSON(200, left)
 	})
 
@@ -283,13 +283,17 @@ func main() {
 	})
 
 	r.POST("/checkPayment", Auth(), func(c *gin.Context) {
-		premium, err := CheckPayment(c.PostForm("orderID"), *c)
+		premium, err := CheckPayment(c.PostForm("orderID"), c)
 
 		if err != nil {
 			c.JSON(200, gin.H{"error": err.Error()})
 		} else {
 			c.JSON(200, gin.H{"premium": premium})
 		}
+	})
+
+	r.GET("/file", func(c *gin.Context) {
+		c.File("C:/OpenServer/domains/dateshipper.api/images/1/avatar.webp")
 	})
 
 	run()
